Pablo merge 1 constructors (menos grupo solar)

+++ src/monopoly/Valor.java
@@ -4,8 +4,8 @@ package monopoly;
 public class Valor {
     //Se incluyen una serie de constantes útiles para no repetir valores.
     public static final float FORTUNA_BANCA = 500000; // Cantidad que tiene inicialmente la Banca
-    public static final double FORTUNA_INICIAL = 9543076.28; // Cantidad que recibe cada jugador al comenzar la partida
-    public static final double SUMA_VUELTA = 1301328.584; // Cantidad que recibe un jugador al pasar pos la Salida
+    public static final float FORTUNA_INICIAL = 9543076.28f; // Cantidad que recibe cada jugador al comenzar la partida
+    public static final float SUMA_VUELTA = 1301328.584f; // Cantidad que recibe un jugador al pasar por la Salida
     
     //Colores del texto:
     public static final String RESET = "\u001B[0m";
